+++ client/firebase.json
@@ -2,6 +2,7 @@
   "hosting": {
     "public": "out",
     "ignore": ["firebase.json", "**/.*", "**/node_modules/**"],
-    "site": "linkzar"
+    "site": "linkzar",
+    "cleanUrls": true
   }
 }
